# server-laravel/openapi.yaml
openapi: 3.0.3
info:
  title: Bluum Patient API
  description: |
    Essential API endpoints for the Bluum patient application.

    **Scope:** This specification documents only the patient-facing mobile/web API endpoints.
    The Nurse Portal API (`/api/nurse/*`) is not included in this specification.

    **Authentication:** All authenticated endpoints use Laravel Sanctum bearer tokens obtained via the `/login` endpoint.

    **Rate Limiting:**
    - `/login` endpoint: Protected by `throttle:auth` middleware
    - All other authenticated endpoints: Protected by `throttle:patient-api` middleware

    **Avatar System:**
    Each patient has an avatar composed of multiple layered images. When a patient object is returned with the `avatar` relationship loaded,
    you'll receive an array of layer images ordered by `layer_number`. To render the avatar:
    1. Load all layer images
    2. Stack them in order (layer 1 = bottom, higher numbers on top)
    3. Composite them together like Photoshop layers

    Available avatars:
    - **Axolotl** (5 layers) - A cute axolotl avatar
    - **Corgi** (7 layers) - A cheerful corgi avatar (default for new patients)

    All avatars are freely available to patients without purchase.
  version: 0.1.0
  contact:
    name: Bluum Development Team
servers:
  - url: http://localhost:8000/api
    description: Local development server
...
paths:
...
  /patient/profile:
    get:
      tags:
        - Profile
      summary: Get patient profile
      description: |
        Retrieve the authenticated patient's full profile with stats.

        **Rate Limiting:** This endpoint is protected by the `throttle:patient-api` middleware.
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Patient profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags:
        - Profile
      summary: Update patient profile
      description: |
        Update the authenticated patient's profile information.

        All fields are optional - only provide the fields you want to update.

        **Rate Limiting:** This endpoint is protected by the `throttle:patient-api` middleware.
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  maxLength: 255
                  description: New username for the patient
                  example: coolpatient123
                avatar_id:
                  type: integer
                  minimum: 1
                  description: ID of the avatar to use (must be at least 1)
                  example: 5
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Profile updated successfully
                  patient:
                    $ref: '#/components/schemas/Patient'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Laravel Sanctum token obtained from the login endpoint.
        Include in the Authorization header as: Bearer {token}
  schemas:
    Patient:
      type: object
      description: Patient model with authentication via pairing code
      required:
        - id
        - username
        - experience
        - gems
      properties:
        id:
          type: integer
          description: Unique identifier for the patient
          example: 1
        username:
          type: string
          description: Patient's username
          maxLength: 255
          example: patient123
        pairing_code:
          type: string
          nullable: true
          description: 6-digit numeric code used for authentication. Null after successful pairing.
          minLength: 6
          maxLength: 6
          pattern: '^\d{6}$'
          example: "123456"
        paired_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when the patient last paired/logged in
          example: '2024-01-20T14:30:00.000000Z'
        avatar_id:
          type: integer
          nullable: true
          description: ID of the avatar selected by the patient
          example: 2
        avatar:
          description: The patient's avatar with all layer images (only present when avatar relationship is loaded)
          allOf:
            - $ref: '#/components/schemas/Avatar'
        experience:
          type: integer
          description: Patient's total XP earned from completing tasks
          minimum: 0
          example: 250
        gems:
          type: integer
          description: Patient's total gems earned from completing tasks
          minimum: 0
          example: 50
        device_identifier:
          type: string
          nullable: true
          description: Unique identifier for the patient's device
          example: "iPhone 14 - iOS 17.1"
        created_at:
          type: string
          format: date-time
          description: Timestamp when the patient record was created
          example: '2024-01-15T10:30:00.000000Z'
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the patient record was last updated
          example: '2024-01-20T14:30:00.000000Z'
    PatientProfile:
      type: object
      description: Extended patient profile with additional statistics
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: patient123
        avatar_id:
          type: integer
          nullable: true
          description: ID of the avatar selected by the patient
          example: 2
        avatar:
          description: The patient's avatar with all layer images (only present when avatar relationship is loaded)
          allOf:
            - $ref: '#/components/schemas/Avatar'
        experience:
          type: integer
          description: Patient's total XP
          example: 250
        gems:
          type: integer
          description: Patient's total gems
          example: 50
        created_at:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00.000000Z'
        stats:
          type: object
          properties:
            total_xp:
              type: integer
              example: 250
            total_gems:
              type: integer
              example: 50
            active_subscriptions:
              type: integer
              description: Number of active task subscriptions
              example: 5
            tasks_completed_today:
              type: integer
              example: 3
            tasks_completed_this_week:
              type: integer
              example: 18
            tasks_completed_total:
              type: integer
              example: 127
            current_streak:
              type: integer
              description: Consecutive days with completed tasks
              example: 7
    Avatar:
      type: object
      description: Avatar model with layered images for compositing
      required:
        - id
        - name
        - base_path
        - layer_count
        - is_default
      properties:
        id:
          type: integer
          description: Unique identifier for the avatar
          example: 2
        name:
          type: string
          description: Name of the avatar type
          example: Corgi
        description:
          type: string
          nullable: true
          description: Description of the avatar
          example: A cheerful corgi avatar
        base_path:
          type: string
          description: Base path to the avatar images directory
          example: assets/Avatar/Corgi
        layer_count:
          type: integer
          description: Total number of layers for this avatar
          example: 7
        is_default:
          type: boolean
          description: Whether this is the default avatar assigned to new patients
          example: true
        layers:
          type: array
          description: Array of layer images ordered by layer_number. Stack these images on top of each other to render the complete avatar (layer 1 = bottom).
          items:
            $ref: '#/components/schemas/AvatarLayer'
    AvatarLayer:
      type: object
      description: Individual layer image for avatar compositing
      required:
        - id
        - layer_number
        - layer_name
        - image_path
      properties:
        id:
          type: integer
          description: Unique identifier for the layer
          example: 6
        layer_number:
          type: integer
          description: Order of the layer (1 is bottom, higher numbers stack on top)
          example: 1
        layer_name:
          type: string
          description: Name of the layer
          example: FirstLayer
        image_path:
          type: string
          description: Full path to the layer image file
          example: assets/Avatar/Corgi/Corgi_FirstLayer.png
...
    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: The given data was invalid.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email:
              - The provided credentials are incorrect.
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unauthenticated.
    TooManyRequests:
      description: Too Many Requests - Rate limit exceeded
      headers:
        Retry-After:
          description: Number of seconds to wait before retrying
          schema:
            type: integer
          example: 60
        X-RateLimit-Limit:
          description: The maximum number of requests allowed in the time window
          schema:
            type: integer
          example: 60
        X-RateLimit-Remaining:
          description: The number of requests remaining in the current time window
          schema:
            type: integer
          example: 0
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Too Many Attempts.
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Server Error